refactor critical CSS code to namespace and subscriber
