+++ inc/classes/optimization/CSS/class-critical-css-generation.php
@@ -1,4 +1,6 @@
 <?php
+namespace WP_Rocket\Optimization\CSS;
+
 defined( 'ABSPATH' ) || die( 'Cheatin&#8217; uh?' );
 
 /**
@@ -9,7 +11,7 @@ defined( 'ABSPATH' ) || die( 'Cheatin&#8217; uh?' );
  *
  * @see WP_Background_Process
  */
-class Rocket_Background_Critical_CSS_Generation extends WP_Background_Process {
+class Critical_CSS_Generation extends \WP_Background_Process {
 	/**
 	 * Process prefix
 	 *
@@ -41,21 +43,7 @@ class Rocket_Background_Critical_CSS_Generation extends WP_Background_Process {
 	 * @access protected
 	 * @var string Critical CSS generator API URL
 	 */
-	protected $critical_css_generator_api_url = 'https://cpcss.wp-rocket.me/api/job/';
-
-	/**
-	 * Launches the background process
-	 *
-	 * @since 2.11
-	 * @author Remy Perona
-	 *
-	 * @access public
-	 * @return array|WP_Error
-	 */
-	public function dispatch() {
-		// Perform remote post.
-		return parent::dispatch();
-	}
+	protected $api_url = 'https://cpcss.wp-rocket.me/api/job/';
 
 	/**
 	 * Perform the optimization corresponding to $item
@@ -69,8 +57,8 @@ class Rocket_Background_Critical_CSS_Generation extends WP_Background_Process {
 	 */
 	protected function task( $item ) {
 		$response = wp_remote_post(
-			$this->critical_css_generator_api_url,
-			array(
+			$this->api_url,
+			[
 				/**
 				 * Filters the parameters sent to the Critical CSS generator API
 				 *
@@ -80,10 +68,14 @@ class Rocket_Background_Critical_CSS_Generation extends WP_Background_Process {
 				 * @param array $params An array of parameters to send to the API.
 				 * @param array $item The item to process.
 				 */
-				'body' => apply_filters( 'rocket_cpcss_job_request', array(
-					'url' => $item['url'],
-				), $item ),
-			)
+				'body' => apply_filters(
+					'rocket_cpcss_job_request',
+					[
+						'url' => $item['url'],
+					],
+					$item
+				),
+			]
 		);
 
 		$transient = get_transient( 'rocket_critical_css_generation_process_running' );
@@ -132,14 +124,15 @@ class Rocket_Background_Critical_CSS_Generation extends WP_Background_Process {
 					rocket_mkdir_p( $critical_css_path );
 				}
 
-				$file_path            = $critical_css_path . '/' . $item['type'] . '.css';
-				$critical_css_content = wp_kses( $job_data->data->critical_path, array( "\'", '\"' ) );
-				$result               = rocket_put_content( $file_path, $job_data->data->critical_path );
+				$file_path     = $critical_css_path . '/' . $item['type'] . '.css';
+				$cpcss_content = wp_kses( $job_data->data->critical_path, [ "\'", '\"' ] );
+				$result        = rocket_put_content( $file_path, $cpcss_content );
 
 				if ( ! $result ) {
 					$transient['items'][] = sprintf(
 						// translators: %1$s = type of content, %2$s = error message.
-						__( 'Critical CSS for %1$s not generated. Error: %2$s', 'rocket' ), $item['type'],
+						__( 'Critical CSS for %1$s not generated. Error: %2$s', 'rocket' ),
+						$item['type'],
 						// translators: %s = critical CSS directory path.
 						sprintf( __( 'The critical CSS content could not be saved as a file in %s.', 'rocket' ), $critical_css_path )
 					);
@@ -162,7 +155,7 @@ class Rocket_Background_Critical_CSS_Generation extends WP_Background_Process {
 	}
 
 	/**
-	 * Gets the returned body of a request to a specific job of the Critical CSS generator API
+	 * Gets the returned body of a request to a specific job from the Critical CSS generator API
 	 *
 	 * @since 2.11
 	 * @author Remy Perona
@@ -172,7 +165,7 @@ class Rocket_Background_Critical_CSS_Generation extends WP_Background_Process {
 	 */
 	protected function get_critical_path( $job_id ) {
 		$response = wp_remote_get(
-			$this->critical_css_generator_api_url . $job_id . '/'
+			$this->api_url . $job_id . '/'
 		);
 
 		return json_decode( wp_remote_retrieve_body( $response ) );
